+++ UnitTests/UnitTest_EditBox/Editing.h
@@ -1,0 +1,6 @@
+#pragma once
+
+void testEditColours();
+void testEditColourBoundaries();
+void testEditKeyboardNavigation();
+void testEditKeyboardCommands();