+++ UnitTests/UnitTest_EditBox/Interaction.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void testEditBoxInteractions();
+void testEditBoxDragAutoscroll();